requirements.txt
# Backend
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
requests
jinja2
email-validator
aiofiles

# OCR + Image Processing
opencv-python
numpy
pillow
pytesseract
pymupdf
fitz
qrcode

# Database
sqlalchemy
databases
bcrypt

# String Matching
rapidfuzz

# ML Models (TrOCR)
transformers
timm
sentencepiece

# (Torch must NOT be installed through requirements.txt)
# torch must be installed using CPU wheel index
